+++ src/main/java/me/gonzager/ex/vehiculos/tiposConduccion/TipoConduccion.java
@@ -1,8 +1,8 @@
 package me.gonzager.ex.vehiculos.tiposConduccion;
 
 public interface TipoConduccion {
-    Double consumo();
-    Double velocidadMaxima();
+    Integer consumo();
+    Integer velocidadMaxima();
     TipoConduccion siguiente();
     TipoConduccion anterior();
 }
